+++ fmt/fmt_str.c
@@ -9,5 +9,5 @@ size_t fmt_str(char *out,const char *in) {
     if (!*t) break; if (s) { *s=*t; ++s; } ++t;
     if (!*t) break; if (s) { *s=*t; ++s; } ++t;
   }
-  return t-in;
+  return (size_t)(t-in);
 }
